noteverse/requirements.txt
django==4.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf_yasg
psycopg2-binary==2.9.6
gunicorn==21.2.0
pillow
pytest
pytest-django